rename tests/models >> tests/base
