+++ tests/loggers/test_base.py
@@ -1,10 +1,10 @@
 import pickle
 from unittest.mock import MagicMock
 
-import tests.models.utils as tutils
+import tests.base.utils as tutils
 from pytorch_lightning import Trainer
 from pytorch_lightning.loggers import LightningLoggerBase, rank_zero_only, LoggerCollection
-from tests.models import LightningTestModel
+from tests.base import LightningTestModel
 
 
 def test_logger_collection():
