+++ tests/loggers/test_comet.py
@@ -5,11 +5,11 @@ from unittest.mock import patch
 import pytest
 import torch
 
-import tests.models.utils as tutils
+import tests.base.utils as tutils
 from pytorch_lightning import Trainer
 from pytorch_lightning.loggers import CometLogger
 from pytorch_lightning.utilities.debugging import MisconfigurationException
-from tests.models import LightningTestModel
+from tests.base import LightningTestModel
 
 
 def test_comet_logger(tmpdir, monkeypatch):
