+++ tests/loggers/test_tensorboard.py
@@ -4,10 +4,10 @@ from argparse import Namespace
 import pytest
 import torch
 
-import tests.models.utils as tutils
+import tests.base.utils as tutils
 from pytorch_lightning import Trainer
 from pytorch_lightning.loggers import TensorBoardLogger
-from tests.models import LightningTestModel
+from tests.base import LightningTestModel
 
 
 def test_tensorboard_logger(tmpdir):
