+++ tests/loggers/test_trains.py
@@ -1,9 +1,9 @@
 import pickle
 
-import tests.models.utils as tutils
+import tests.base.utils as tutils
 from pytorch_lightning import Trainer
 from pytorch_lightning.loggers import TrainsLogger
-from tests.models import LightningTestModel
+from tests.base import LightningTestModel
 
 
 def test_trains_logger(tmpdir):
