+++ tests/test_cpu_models.py
@@ -4,12 +4,12 @@ import warnings
 import pytest
 import torch
 
-import tests.models.utils as tutils
+import tests.base.utils as tutils
 from pytorch_lightning import Trainer
 from pytorch_lightning.callbacks import (
     EarlyStopping,
 )
-from tests.models import (
+from tests.base import (
     TestModelBase,
     LightTrainDataloader,
     LightningTestModel,
