+++ tests/test_gpu_models.py
@@ -3,7 +3,7 @@ import os
 import pytest
 import torch
 
-import tests.models.utils as tutils
+import tests.base.utils as tutils
 from pytorch_lightning import Trainer
 from pytorch_lightning.callbacks import ModelCheckpoint
 from pytorch_lightning.core import memory
@@ -12,7 +12,7 @@ from pytorch_lightning.trainer.distrib_parts import (
     determine_root_gpu_device,
 )
 from pytorch_lightning.utilities.debugging import MisconfigurationException
-from tests.models import LightningTestModel
+from tests.base import LightningTestModel
 
 PRETEND_N_OF_GPUS = 16
